IndexedDB/tests/submissions/Microsoft/idbobjectstore_get.htm

<!DOCTYPE html>
<html>
<head>
    <meta content="text/html; charset=utf-8" http-equiv="content-type" />
    <title> IDBObjectStore.get() - key is a number </title>
    <script type="text/javascript" src="support.js"></script>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript">
        var objectStoreName = "objectstore";
        var record = {key:3.14159265, property:"data"};
        var db = null;
        
        var t = async_test();
        
        function RunTest() {
            var rqOpen = window.indexedDB.open(databaseName, databaseVersion);
            rqOpen.onupgradeneeded = t.step_func(function(event) {
                db = event.target.result;
                db.onerror = t.step_func(assert_database_error);
                var objStore = db.createObjectStore(objectStoreName, {keyPath:"key"});
                objStore.add(record);
            });
            rqOpen.onerror = t.step_func(assert_open_request_error);
            rqOpen.onsuccess = t.step_func(GetRecord);
        }
        
        function GetRecord() {
            var txn = db.transaction(objectStoreName, "readonly");
            var objStore = txn.objectStore(objectStoreName);
            var rqGet = objStore.get(record.key);
            rqGet.onsuccess = t.step_func(function(event) {
                assert_equals(event.target.result.key, record.key, 'event.target.result.key');
                assert_equals(event.target.result.property, record.property, 'event.target.result.property');
                t.done();
            });
        }
        
        add_completion_callback(function() { if(db) db.close(); });
        
        setup(function() {
            var rqDelete = window.indexedDB.deleteDatabase(databaseName);
            rqDelete.onsuccess = t.step_func( RunTest );
        });
    </script>
</head>
<body>
    <div id="log"></div>
</body>
</html>
